Basic UI Test in SDK
